backend/requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
python-multipart